Move RTree implementation to impl
