+++ src/mesh/RTree.hpp
@@ -73,69 +73,11 @@ std::ostream& operator<<(std::ostream& out, PrimitiveIndex val);
 /// The axis aligned bounding box based on the Vertex Type
 using AABB = boost::geometry::model::box<Eigen::VectorXd>;
 
-/** Holding a reference to a Mesh it is a functor for packing
- *  PrimitiveIndex into AABBs.
- */
-class AABBGenerator
-{
-public:
-  /// Constructs a generator for a given mesh
-  explicit AABBGenerator(const mesh::Mesh &mesh)
-      : mesh_(mesh){};
-
-  /** constructs a AABB of a PrimitiveIndex.
-   *
-   * This operator takes a PrimitiveIndex and fetches the actual primitive from the mesh.
-   * It then constructs an AABB using boost::geometry::return_envelope() and returns it.
-   *
-   * \param pi the index to build the AABB for
-   * \returns a AABB for the primitive
-   */
-  AABB operator()(const PrimitiveIndex &pi) const
-  {
-    switch (pi.type) {
-    case (Primitive::Vertex):
-      return boost::geometry::return_envelope<AABB>(mesh_.vertices()[pi.index]);
-    case (Primitive::Edge):
-      return boost::geometry::return_envelope<AABB>(mesh_.edges()[pi.index]);
-    case (Primitive::Triangle):
-      return boost::geometry::return_envelope<AABB>(mesh_.triangles()[pi.index]);
-    case (Primitive::Quad):
-      return boost::geometry::return_envelope<AABB>(mesh_.quads()[pi.index]);
-    default:
-      assertion(false, "AABB generation for this Primitive is not implemented!")
-    }
-  }
-
-private:
-  /// The mesh used look the primitives up
-  const mesh::Mesh &mesh_;
-};
-
 /// The rtree capable of indexing primitives of an entire Mesh
 using PrimitiveRTree = boost::geometry::index::rtree<std::pair<AABB, PrimitiveIndex>, boost::geometry::index::rstar<16>>;
 
 /// The shared_ptr convenience type for PrimitiveRTree
 using PtrPrimitiveRTree = std::shared_ptr<PrimitiveRTree>;
-
-/** indexes a container of primitives into an rtree.
- *
- * The algorithm indexes every primitive of the given container using the passed generator.
- * It inserts its result and the PrimitiveIndex into the rtree as a std::pair.
- *
- * \param[inout] rtree the rtree to index into
- * \param gen the Generator generating something to index rtree::value_type.
- * \param conti the Container to index
- */
-template <typename Container, typename Generator = AABBGenerator>
-void indexPrimitive(PrimitiveRTree &rtree, const Generator& gen, const Container &conti)
-{
-  using ValueType = typename std::remove_reference<typename std::remove_cv<typename Container::value_type>::type>::type;
-  for (size_t i = 0; i < conti.size(); ++i) {
-    PrimitiveIndex index{as_primitive_enum<ValueType>::value, i};
-    rtree.insert(std::make_pair(gen(index), index));
-  }
-}
 
 /** Indexes a given mesh and returns a PrimitiveRTree holding the index
  *
@@ -145,16 +87,7 @@ void indexPrimitive(PrimitiveRTree &rtree, const Generator& gen, const Container
  *
  * \return the tree containing the indexed primitives descibed above
  */
-inline PrimitiveRTree indexMesh(const Mesh &mesh)
-{
-  AABBGenerator  gen{mesh};
-  PrimitiveRTree tree;
-  indexPrimitive(tree, gen, mesh.vertices());
-  indexPrimitive(tree, gen, mesh.edges());
-  indexPrimitive(tree, gen, mesh.triangles());
-  indexPrimitive(tree, gen, mesh.quads());
-  return tree;
-}
+PrimitiveRTree indexMesh(const Mesh &mesh);
 
 class rtree {
 public:
